Refactor a few grunt tasks to the tasks folder
